+++ zetaclient/chains/evm/signer/signer_test.go
@@ -145,7 +145,12 @@ func TestSigner_TryProcessOutbound(t *testing.T) {
 		// ARRANGE
 		// Setup evm signer
 		evmSigner := newTestSuite(t)
+
+		// Setup txData struct
 		cctx := getCCTX(t)
+		nonce := cctx.GetCurrentOutboundParam().TssNonce
+		txData, _, err := NewOutboundData(ctx, cctx, zerolog.Logger{})
+		require.NoError(t, err)
 
 		// Test with mock client that has keys
 		client := mocks.NewZetacoreClient(t).
@@ -155,8 +160,11 @@ func TestSigner_TryProcessOutbound(t *testing.T) {
 		zetaRepo := zrepo.New(client, chains.Ethereum, mode.StandardMode)
 
 		// mock evm client "NonceAt"
-		nonce := uint64(123)
 		evmSigner.evmServer.MockNonceAt(nonce)
+
+		// Mock the digest to be signed
+		digest := getGasWithdrawDigest(t, evmSigner.Signer, txData)
+		mockSignature(t, evmSigner.Signer, nonce, digest)
 
 		// ACT
 		evmSigner.TryProcessOutbound(ctx, cctx, zetaRepo, nonce)
@@ -201,10 +209,14 @@ func TestSigner_BroadcastOutbound(t *testing.T) {
 
 	// Setup txData struct
 	cctx := getCCTX(t)
-	nonce := uint64(123)
+	nonce := cctx.GetCurrentOutboundParam().TssNonce
 	txData, skip, err := NewOutboundData(ctx, cctx, zerolog.Logger{})
 	require.NoError(t, err)
 	require.False(t, skip)
+
+	// Mock the digest to be signed
+	digest := getERC20WithdrawDigest(t, evmSigner.Signer, txData)
+	mockSignature(t, evmSigner.Signer, nonce, digest)
 
 	// Mock evm client "NonceAt"
 	evmSigner.evmServer.MockNonceAt(nonce)