+++ zetaclient/testutils/mocks/tss.go
@@ -110,6 +110,10 @@ func (tss *TSS) SignBatch(_ context.Context, digests [][]byte, _, _ uint64, _ in
 	return sigs, nil
 }
 
+func (tss *TSS) IsSignatureCached(chainID int64, digests [][]byte) bool {
+	return false
+}
+
 func (tss *TSS) UpdatePrivateKey(pk *ecdsa.PrivateKey) {
 	tss.privateKey = pk
 }