+++ changelog.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+* [4427](https://github.com/zeta-chain/node/pull/4427) - improve EVM chain outbound performance with batch and sequential keysign
+
 ### Tests
 
 ### Fixes